refactor: moved cards and ember effect into ui folder in components
