+++ app/frontend/src/components/admin/approvalCard.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Card from '../Card';
+import Card from '../ui/Card';
 
 interface RoleRequest {
   id: string;
