+++ app/frontend/src/components/ui/EmberEffect.tsx
