+++ app/frontend/src/pages/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import Image from 'next/image';
-import EmberField from '../components/EmberEffect';
+import EmberField from '../components/ui/EmberEffect';
 
 export default function Landing() {
   const router = useRouter();
